update function to print binary

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -10,7 +10,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int i, flag;
+	int i, flag;
 
 	if (n == 0)
 		_putchar('0');
